Base Logger for loggers
