+++ Logging/Logger.cs
@@ -74,106 +74,27 @@ namespace Logging
         Debug
     }
 
-    /// <summary>
-    /// All user event logger handler
-    /// </summary>
-    public class Logger
+    public class ConsoleLogger : LoggerBase
     {
-        /// <summary>
-        /// Class that handles the writing depending on the Logging Mode provided
-        /// </summary>
-        private IWriter messageWriter;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="Logger"/> class with the specified logging mode format
-        /// </summary>
-        /// <param name="filePath">Path where the file will be stored</param>
-        /// <param name="fileName">Name of the file</param>
-        /// <param name="loggingMode">Logging mode that will be used</param>
-        /// <param name="writingMode">Used writing mode</param>
-        public Logger(string filePath, string fileName, LoggingMode loggingMode, WritingMode writingMode = WritingMode.Appending)
+        public ConsoleLogger()
         {
-            this.LoggingMode = loggingMode;
-            this.FilePath = filePath;
-            this.FileName = fileName;
-            this.WritingMode = writingMode;
-            this.InstantiateWriters();
+            this.messageWriter = new ConsoleWriter();
         }
+    }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="Logger"/> class with a logging mode with any specified mode
-        /// </summary>
-        /// <param name="loggingMode">Logging mode that will be used</param>
-        public Logger(LoggingMode loggingMode)
+    public class DebugLogger : LoggerBase
+    {
+        public DebugLogger()
         {
-            this.LoggingMode = loggingMode;
-            this.InstantiateWriters();
+            this.messageWriter = new DebuggerWriter();
         }
+    }
 
-        /// <summary>
-        /// Gets or sets the properties that the messages have
-        /// </summary>
-        public MessageProperties MessageProperties { get; set; } = new MessageProperties();
-
-        /// <summary>
-        /// Gets or sets the properties that each message level has
-        /// </summary>
-        public MessageLevelProperties MessageLevelProperties { get; set; } = new MessageLevelProperties();
-
-        /// <summary>
-        /// Gets the file path where the logging files will be stored, has to be set with the override constructor
-        /// </summary>
-        public string FilePath { get; private set; }
-
-        /// <summary>
-        /// Gets the fileName that will be used to store the logging, has to be set with the override constructor
-        /// </summary>
-        public string FileName { get; private set; }
-
-        /// <summary>
-        /// Gets the usage mode, it cannot be modified after instantiation
-        /// </summary>
-        public LoggingMode LoggingMode { get; private set; }
-
-        /// <summary>
-        /// Gets the writing mode, it cannot be modified after instantiation, has to be set with the override constructor
-        /// </summary>
-        public WritingMode WritingMode { get; private set; }
-
-        /// <summary>
-        /// Writes the message depending on the specified message level and the logging mode set in the logger class
-        /// </summary>
-        /// <param name="msg">Message to log</param>
-        /// <param name="messageLevel">Indicates the level of the message</param>
-        public void WriteMessage(string msg, MessageLevel messageLevel = MessageLevel.Info)
+    public class FileLogger : LoggerBase
+    {
+        public FileLogger(string filePath, string fileName, WritingMode writingMode = WritingMode.Appending)
         {
-            msg = MessageBuilder.MessageStringBuilder(msg, messageLevel, this.MessageProperties, this.MessageLevelProperties);
-            MessageLevelProperties.IncreaseCounter(messageLevel);
-            this.messageWriter.WriteMessage(msg, messageLevel, this.MessageProperties, this.MessageLevelProperties);
-        }
-
-        /// <summary>
-        /// This class initializes the Message Write depending on the provided logging mode
-        /// </summary>
-        private void InstantiateWriters()
-        {
-            switch (LoggingMode)
-            {
-                case LoggingMode.Console:
-                    this.messageWriter = new ConsoleWriter();
-                    break;
-
-                case LoggingMode.Debug:
-                    this.messageWriter = new DebuggerWriter();
-                    break;
-
-                case LoggingMode.TextFile:
-                    this.messageWriter = new TextFileWriter(this.FilePath, this.FileName, this.WritingMode);
-                    break;
-
-                default:
-                    break;
-            }
+            this.messageWriter = new TextFileWriter(filePath, fileName, writingMode);
         }
     }
 }
