+++ UnitTestLogging/UnitTestLogger.cs
@@ -6,7 +6,7 @@ namespace UnitTestLogging
     [TestClass]
     public class UnitTestLoggerCounter
     {
-        private readonly Logger logger = new Logger(LoggingMode.Console);
+        private readonly ConsoleLogger logger = new ConsoleLogger();
         private readonly string sampleText = "Text";
 
         [TestMethod]
